split V2 tests into a separate file
